Additional controller and service methods

+++ PartyTimeREST/src/main/java/com/pasciak/partytime/repositories/UserRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pasciak.partytime.entities.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 	User findByUsername(String username);
 	
-	User findById(int id);
+	User findById(long id);
 	
 	List<User> findAll();
